requirements.txt
Django==5.0.4
djangorestframework==3.15.1
celery==5.4.0
django-redis==5.4.0
openpyxl==3.1.2